requirements.txt
pillow=10.2.0
PyPDF2==3.0.1
opencv-python==4.6.0.66
glob2==0.7
img2pdf==0.5.1